refactor(shopkeeper): use singleton pattern and register routes/listeners explicitly
- Replace static accessors with Shopkeeper.$instance singleton
- Rename registerRoute/registerListeners to registerRoutes/registerWebhookListeners
- Fix lazy listener registration with [import, 'handle'] tuple format
- Fix middleware DI by using value imports instead of type imports
- Check webhook secret before signature validation in middleware
- Add reflect-metadata import in test setup
