--- packages/shopkeeper/providers/README.md
@@ -1,5 +1,0 @@
-# The providers directory
-
-The `providers` directory contains the service providers exported by your application. Make sure to register these providers within the `exports` collection (aka package entrypoints) defined within the `package.json` file.
-
-Learn more about [package entrypoints](https://nodejs.org/api/packages.html#package-entry-points).
